# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SportsLine API",
    "version": "1.0.0",
    "description": "API REST para la empresa ficticia SportsLine."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nombre": {
            "type": "string"
          },
          "codigo": {
            "type": "string"
          },
          "precio": {
            "type": "number",
            "format": "float"
          },
          "categoria": {
            "type": "string"
          },
          "stock": {
            "type": "integer"
          }
        }
      },
      "ProductCreateRequest": {
        "type": "object",
        "required": [
          "nombre",
          "codigo",
          "precio",
          "categoria",
          "stock"
        ],
        "properties": {
          "nombre": {
            "type": "string",
            "example": "Balón de baloncesto"
          },
          "codigo": {
            "type": "string",
            "example": "BAL-999"
          },
          "precio": {
            "type": "number",
            "format": "float",
            "example": 49.99
          },
          "categoria": {
            "type": "string",
            "example": "Baloncesto"
          },
          "stock": {
            "type": "integer",
            "example": 25
          }
        }
      },
...
      "OrderItem": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "quantity": {
            "type": "integer"
          }
        }
      },
      "Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "clienteId": {
            "type": "string",
            "format": "uuid"
          },
          "usuarioId": {
            "type": "string",
            "format": "uuid"
          },
          "fecha": {
            "type": "string",
            "format": "date-time"
          },
          "estado": {
            "type": "string",
            "enum": [
              "pendiente",
              "preparando",
              "entregado"
            ]
          },
          "total": {
            "type": "number",
            "format": "float"
          }
        }
      },
      "OrderCreateRequest": {
        "type": "object",
        "required": [
          "clienteId",
          "items"
        ],
        "properties": {
          "clienteId": {
            "type": "string",
            "format": "uuid",
            "example": "7e8f3c3d-1dcb-4c54-9f4f-53e280d1a0f2"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            },
            "example": [
              {
                "productId": "b3ab65cf-63d1-4d9a-9c80-15d2ac3d5c60",
                "quantity": 2
              },
              {
                "productId": "d9a85932-80d6-4feb-b034-4520fdfa3a8b",
                "quantity": 1
              }
            ]
          },
          "sensitiveData": {
            "type": "object",
            "properties": {
              "notasInternas": {
                "type": "string",
                "example": "Entregar antes de las 10am"
              },
              "pagoReferencia": {
                "type": "string",
                "example": "TRANS-987654"
              }
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object"
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/orders": {
      "get": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Pedidos"
        ],
        "summary": "Lista pedidos con filtros opcionales",
        "parameters": [
          {
            "in": "query",
            "name": "clienteId",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "productoId",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "estado",
            "schema": {
              "type": "string",
              "enum": [
                "pendiente",
                "preparando",
                "entregado"
              ]
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de pedidos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "orders": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Order"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Pedidos"
        ],
        "summary": "Crea un pedido",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pedido creado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "order": {
                      "$ref": "#/components/schemas/Order"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Stock insuficiente o datos inválidos"
          },
          "404": {
            "description": "Producto o cliente no encontrado"
          }
        }
      }
    },
    "/api/v1/orders/{id}": {
      "get": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Pedidos"
        ],
        "summary": "Obtiene un pedido por ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pedido encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "order": {
                      "$ref": "#/components/schemas/Order"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Pedido no encontrado"
          }
        }
      }
    },
    "/api/v1/orders/{id}/estado": {
      "patch": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Pedidos"
        ],
        "summary": "Actualiza el estado de un pedido",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "estado": {
                    "type": "string",
                    "enum": [
                      "pendiente",
                      "preparando",
                      "entregado"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pedido actualizado"
          },
          "400": {
            "description": "Estado inválido"
          },
          "404": {
            "description": "Pedido no encontrado"
          }
        }
      }
    },
    "/api/v1/orders/{id}/sensible": {
      "get": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Pedidos"
        ],
        "summary": "Obtiene los datos sensibles cifrados de un pedido",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Datos sensibles descifrados",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Permisos insuficientes"
          },
          "404": {
            "description": "Pedido no encontrado o sin datos sensibles"
          }
        }
      }
    },
    "/api/v1/products": {
      "get": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Productos"
        ],
        "summary": "Lista todos los productos",
        "responses": {
          "200": {
            "description": "Listado de productos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "products": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Product"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error interno"
          }
        }
      },
      "post": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Productos"
        ],
        "summary": "Crea un nuevo producto",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Producto creado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "product": {
                      "$ref": "#/components/schemas/Product"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos"
          },
          "401": {
            "description": "No autorizado"
          },
          "409": {
            "description": "Código duplicado"
          }
        }
      }
    },
    "/api/v1/products/{id}": {
      "get": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Productos"
        ],
        "summary": "Obtiene un producto por ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Producto encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "product": {
                      "$ref": "#/components/schemas/Product"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "Producto no encontrado"
          }
        }
      },
      "put": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Productos"
        ],
        "summary": "Actualiza un producto existente",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Producto actualizado"
          },
          "404": {
            "description": "Producto no encontrado"
          },
          "409": {
            "description": "Código duplicado"
          }
        }
      },
      "delete": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Productos"
        ],
        "summary": "Elimina un producto",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Producto eliminado"
          },
          "404": {
            "description": "Producto no encontrado"
          }
        }
      }
    }
  },
...
}